Fix GPS signal/slot connection
QObject::connect: No such signal QGPSDDevice::updatePosition(float, float, QDateTime, float, float, float) in ../../src/MainWindow.cpp:4000
QObject::connect:  (receiver name: 'MainWindow')

+++ src/MainWindow.cpp
@@ -3994,8 +3994,8 @@ void MainWindow::on_gpsConnectAction_triggered()
     QGPSS60Device* aGps = new QGPSS60Device();
 #endif
     if (aGps->openDevice()) {
-        connect(aGps, SIGNAL(updatePosition(float, float, QDateTime, float, float, float)),
-            this, SLOT(updateGpsPosition(float, float, QDateTime, float, float, float)));
+        connect(aGps, SIGNAL(updatePosition(qreal, qreal, QDateTime, qreal, qreal, qreal)),
+            this, SLOT(updateGpsPosition(qreal, qreal, QDateTime, qreal, qreal, qreal)));
 
         ui->gpsConnectAction->setEnabled(false);
         ui->gpsReplayAction->setEnabled(false);
@@ -4049,8 +4049,8 @@ void MainWindow::on_gpsDisconnectAction_triggered()
     ui->gpsRecordAction->setChecked(false);
     ui->gpsPauseAction->setChecked(false);
 
-    disconnect(theGPS->getGpsDevice(), SIGNAL(updatePosition(float, float, QDateTime, float, float, float)),
-        this, SLOT(updateGpsPosition(float, float, QDateTime, float, float, float)));
+    disconnect(theGPS->getGpsDevice(), SIGNAL(updatePosition(qreal, qreal, QDateTime, qreal, qreal, qreal)),
+        this, SLOT(updateGpsPosition(qreal, qreal, QDateTime, qreal, qreal, qreal)));
     theGPS->stopGps();
     theGPS->resetGpsStatus();
 }
